Create export tile service
