+++ frontend/src/app/modules/home/dashboard/tiles/top-active-issues-tile/top-active-issues-tile.component.ts
@@ -17,6 +17,8 @@ import { jsPDF } from 'jspdf';
 import * as FileSaver from 'file-saver';
 import 'blob';
 import { ExportType } from '@shared/models/tile/enums/export-type';
+import { ExportTileService } from '@core/services/export-tile.service';
+import { SpinnerService } from '@core/services/spinner.service';
 
 const JsPDF = jsPDF;
 
@@ -42,6 +44,8 @@ export class TopActiveIssuesTileComponent extends BaseComponent implements OnIni
         private tileDialogService: TileDialogService,
         private issueService: IssueService,
         private authService: AuthenticationService,
+        private exportTileService: ExportTileService,
+        private spinner: SpinnerService
     ) {
         super();
     }
@@ -66,36 +70,9 @@ export class TopActiveIssuesTileComponent extends BaseComponent implements OnIni
     }
 
     exportTile(exportType: ExportType) {
-        const data = this.data.host.nativeElement;
-        const smallPdf = 300;
-        const mediumPdf = 440;
-        const bigPdf = 880;
-        html2canvas(data).then(canvas => {
-            if (exportType === ExportType.Jpg) {
-                canvas.toBlob((blob) => {
-                    FileSaver.saveAs(blob, `${new Date().toLocaleString().replace(':', '_')}.jpg`);
-                });
-            }
-            if (exportType === ExportType.Png) {
-                canvas.toBlob((blob) => {
-                    FileSaver.saveAs(blob, `${new Date().toLocaleString().replace(':', '_')}.png`);
-                });
-            }
-            if (exportType === ExportType.Pdf) {
-                const contentDataURL = canvas.toDataURL('image/png', 1.0);
-                const pdf = new JsPDF('p', 'mm', 'a4');
-                if (data.clientWidth >= smallPdf && data.clientWidth <= mediumPdf) {
-                    pdf.addImage(contentDataURL, 'PNG', 60, 10, canvas.height / 6, 0);
-                }
-                if (data.clientWidth >= mediumPdf && data.clientWidth <= bigPdf) {
-                    pdf.addImage(contentDataURL, 'PNG', 35, 10, canvas.height / 4, 0);
-                }
-                if (data.clientWidth >= bigPdf) {
-                    pdf.addImage(contentDataURL, 'PNG', 0, 10, canvas.height / 2.7, 0);
-                }
-                pdf.save(`${new Date().toLocaleString().replace(':', '_')}.pdf`);
-            }
-        });
+        this.spinner.show(true);
+        this.exportTileService.exportTile(exportType, this.data.host.nativeElement);
+        this.spinner.hide();
     }
 
     private getTileSettings() {
